pyproject.toml
[tool.poetry]
name = "nonebot-plugin-dg-lab-play"
version = "1.0.0"
description = "nonebot2 机器人插件 - 在群里和大家一起玩郊狼吧！"
authors = ["Ljzd-PRO <ljzd@office.ljzd-pro.asia>"]
readme = "README.md"
homepage = "https://github.com/Ljzd-PRO/nonebot-plugin-dg-lab-play"
repository = "https://github.com/Ljzd-PRO/nonebot-plugin-dg-lab-play"
documentation = "https://github.com/Ljzd-PRO/nonebot-plugin-dg-lab-play"

keywords = ["nonebot", "nonebot2", "dg-lab", "dg-lab-v3"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.2.0"
pydglab-ws = "^1.0.0"
nonebot-plugin-send-anything-anywhere = "^0.6.1"
qrcode = {extras = ["pil"], version = "^7.4.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
